pyproject.toml
[project]
name = "gemini-flash-tts-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.45.0",
    "pydub>=0.25.1",
]
